extract I/O expander drivers
Move PCA9555 and PCF8574/8575 register access out of the platform GPIO module into a dedicated drivers component. Preserve the existing GPIO compatibility surface and hardware behavior while wiring the new directory into Linux and PlatformIO builds.
