+++ Makefile
@@ -6,8 +6,8 @@ LD=$(CXX)
 LIBS=pthread mosquitto ssl crypto i2c lgpio
 LDFLAGS=$(addprefix -l,$(LIBS))
 BINARY=OpenSprinkler
-SOURCES=main.cpp OpenSprinkler.cpp notifier.cpp program.cpp opensprinkler_server.cpp utils.cpp weather.cpp gpio.cpp mqtt.cpp smtp.c RCSwitch.cpp i2cd.cpp ads1115.cpp $(wildcard boards/*.cpp) $(wildcard sensors/*.cpp) $(wildcard external/TinyWebsockets/tiny_websockets_lib/src/*.cpp) $(wildcard external/OpenThings-Framework-Firmware-Library/*.cpp)
-HEADERS=$(wildcard *.h) $(wildcard *.hpp) $(wildcard boards/*.h) $(wildcard sensors/*.h)
+SOURCES=main.cpp OpenSprinkler.cpp notifier.cpp program.cpp opensprinkler_server.cpp utils.cpp weather.cpp gpio.cpp mqtt.cpp smtp.c RCSwitch.cpp i2cd.cpp ads1115.cpp $(wildcard boards/*.cpp) $(wildcard drivers/*.cpp) $(wildcard sensors/*.cpp) $(wildcard external/TinyWebsockets/tiny_websockets_lib/src/*.cpp) $(wildcard external/OpenThings-Framework-Firmware-Library/*.cpp)
+HEADERS=$(wildcard *.h) $(wildcard *.hpp) $(wildcard boards/*.h) $(wildcard drivers/*.h) $(wildcard sensors/*.h)
 OBJECTS=$(addsuffix .o,$(basename $(SOURCES)))
 
 .PHONY: all
