+++ gpio.cpp
@@ -25,118 +25,7 @@
 
 #if defined(ARDUINO)
 
-#include <Wire.h>
 #include "defines.h"
-
-unsigned char IOEXP::detectType(uint8_t address) {
-	Wire.beginTransmission(address);
-	if(Wire.endTransmission()!=0) return IOEXP_TYPE_NONEXIST; // this I2C address does not exist
-
-	Wire.beginTransmission(address);
-	Wire.write(NXP_INVERT_REG); // ask for polarity register
-	Wire.endTransmission();
-
-	if(Wire.requestFrom(address, (uint8_t)2) != 2) return IOEXP_TYPE_UNKNOWN;
-	uint8_t low = Wire.read();
-	uint8_t high = Wire.read();
-	if(low==0x00 && high==0x00) {
-		return IOEXP_TYPE_9555; // PCA9555 has polarity register which inits to 0
-	}
-	return IOEXP_TYPE_8575;
-}
-
-void PCA9555::pinMode(uint8_t pin, uint8_t IOMode) {
-	uint16_t config = i2c_read(NXP_CONFIG_REG);
-	if(IOMode == OUTPUT) {
-		config &= ~(1 << pin); // config bit set to 0 for output pin
-	} else {
-		config |= (1 << pin);  // config bit set to 1 for input pin
-	}
-	i2c_write(NXP_CONFIG_REG, config);
-}
-
-uint16_t PCA9555::i2c_read(uint8_t reg) {
-	if(address==255)	return 0xFFFF;
-	Wire.beginTransmission(address);
-	Wire.write(reg);
-	Wire.endTransmission();
-	if(Wire.requestFrom(address, (uint8_t)2) != 2) {return 0xFFFF; DEBUG_PRINTLN("GPIO error");}
-	uint16_t data0 = Wire.read();
-	uint16_t data1 = Wire.read();
-	return data0+(data1<<8);
-}
-
-void PCA9555::i2c_write(uint8_t reg, uint16_t v){
-	if(address==255)	return;
-	Wire.beginTransmission(address);
-	Wire.write(reg);
-	Wire.write(v&0xff);
-	Wire.write(v>>8);
-	Wire.endTransmission();
-}
-
-void PCA9555::shift_out(uint8_t plat, uint8_t pclk, uint8_t pdat, uint8_t v) {
-	if(plat<IOEXP_PIN || pclk<IOEXP_PIN || pdat<IOEXP_PIN)
-		return; // the definition of each pin must be offset by IOEXP_PIN to begin with
-
-	plat-=IOEXP_PIN;
-	pclk-=IOEXP_PIN;
-	pdat-=IOEXP_PIN;
-
-	uint16_t output = i2c_read(NXP_OUTPUT_REG); // keep a copy of the current output registers
-
-	output &= ~(1<<plat); i2c_write(NXP_OUTPUT_REG, output); // set latch low
-
-	for(uint8_t s=0;s<8;s++) {
-		output &= ~(1<<pclk); i2c_write(NXP_OUTPUT_REG, output); // set clock low
-
-		if(v & ((unsigned char)1<<(7-s))) {
-			output |= (1<<pdat);
-		} else {
-			output &= ~(1<<pdat);
-		}
-		i2c_write(NXP_OUTPUT_REG, output); // set data pin according to bits in v
-
-		output |= (1<<pclk); i2c_write(NXP_OUTPUT_REG, output); // set clock high
-	}
-
-	output |= (1<<plat); i2c_write(NXP_OUTPUT_REG, output); // set latch high
-}
-
-uint16_t PCF8575::i2c_read(uint8_t reg) {
-	if(address==255)	return 0xFFFF;
-	Wire.beginTransmission(address);
-	if(Wire.requestFrom(address, (uint8_t)2) != 2) return 0xFFFF;
-	uint16_t data0 = Wire.read();
-	uint16_t data1 = Wire.read();
-	Wire.endTransmission();
-	return data0+(data1<<8);
-}
-
-void PCF8575::i2c_write(uint8_t reg, uint16_t v) {
-	if(address==255)	return;
-	Wire.beginTransmission(address);
-	// todo: handle inputmask (not necessary unless if using any pin as input)
-	Wire.write(v&0xff);
-	Wire.write(v>>8);
-	Wire.endTransmission();
-}
-
-uint16_t PCF8574::i2c_read(uint8_t reg) {
-	if(address==255)	return 0xFFFF;
-	Wire.beginTransmission(address);
-	if(Wire.requestFrom(address, (uint8_t)1) != 1) return 0xFFFF;
-	uint16_t data = Wire.read();
-	Wire.endTransmission();
-	return data;
-}
-
-void PCF8574::i2c_write(uint8_t reg, uint16_t v) {
-	if(address==255)	return;
-	Wire.beginTransmission(address);
-	Wire.write((uint8_t)(v&0xFF) | inputmask);
-	Wire.endTransmission();
-}
 
 #include "OpenSprinkler.h"
 
@@ -283,4 +172,3 @@ void digitalWrite(int pin, unsigned char value) {}
 unsigned char digitalRead(int pin) {return 0;}
 
 #endif
-
