+++ platformio.ini
@@ -24,7 +24,7 @@ lib_deps =
 	./external/OpenThings-Framework-Firmware-Library
 	;https://github.com/OpenThingsIO/OpenThings-Framework-Firmware-Library @ ^0.2.0
 ; ignore html2raw.cpp source file for firmware compilation (external helper program)
-build_src_filter = +<*> +<boards/*> +<sensors/*> -<html/*> -<tests/*> --<external/*>
+build_src_filter = +<*> +<boards/*> +<drivers/*> +<sensors/*> -<html/*> -<tests/*> --<external/*>
 upload_speed = 460800
 monitor_speed = 115200
 board_build.flash_mode = dio
